Move binaries into cmd/ where they belong
Also fix egressip.go so it compiles on Mac correctly
