+++ cmd/sdn-cni-plugin/openshift-sdn.go
@@ -1,0 +1,5 @@
+// +build !linux
+
+package main
+
+func main() {}
