+++ pkg/network/node/egressip.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package node
 
 import (
